% Auto-generated: do not edit by hand
\name{''DashFloorPlan}

\alias{''DashFloorPlan}

\title{DashFloorPlan component}

\description{

}

\usage{
''DashFloorPlan(id=NULL, selection=NULL, update=NULL, height=NULL, width=NULL, shapes=NULL,
data=NULL, image=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{selection}{Logical | numeric | character | named list | unnamed list. Currently selected polygon.}

\item{update}{Logical. Flag to trigger update from Dash.}

\item{height}{Numeric. Height of the svg panel in pixels.}

\item{width}{Numeric. Width of the svg panel in pixels.}

\item{shapes}{Unnamed list. Array of polygons to display over image.}

\item{data}{Named list. Dict of data to associated with the polygons.}

\item{image}{Character. Path to image file.}
}
