% Auto-generated: do not edit by hand
\name{''DashFloorPlan}

\alias{''DashFloorPlan}

\title{DashFloorPlan component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashFloorPlan(id=NULL, height=NULL, width=NULL, data=NULL, image=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{height}{Numeric. Height of the svg panel in pixels.}

\item{width}{Numeric. Width of the svg panel in pixels.}

\item{data}{Unnamed list. Array of polygons to display over image.}

\item{image}{Character. Path to image file.}
}
